attempt to fix a few failing tests
