+++ .github/actions/get_exclude_dirs/action.yml
@@ -42,8 +42,10 @@ runs:
               examples/experimental/test/end_to_end/http_server/ethers_base
               examples/experimental/test/end_to_end/http_server/http_outcall_fetch
               examples/experimental/test/end_to_end/http_server/ic_evm_rpc
+              examples/experimental/test/end_to_end/http_server/nest
               examples/experimental/test/end_to_end/http_server/open_value_sharing
               examples/experimental/test/end_to_end/http_server/sqlite_drizzle
+              examples/experimental/test/end_to_end/http_server/sqlite_typeorm
               ') }}"
 
               SLOW_EXCLUSIONS="${{ format('
