+++ examples/experimental/test/end_to_end/http_server/express/src/frontend/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../../../../../../../../tsconfig.dev.json",
+    "compilerOptions": {
+        "experimentalDecorators": true
+    }
+}
