+++ examples/experimental/test/end_to_end/http_server/fetch_ic/src/frontend/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../../../../../../../tsconfig.dev.json",
+    "compilerOptions": {
+        "allowSyntheticDefaultImports": true,
+        "experimentalDecorators": true
+    }
+}
