+++ examples/experimental/test/end_to_end/http_server/fetch_ic/tsconfig.json
@@ -1,7 +1,3 @@
 {
-    "extends": "../../../../../../tsconfig.dev.json",
-    "compilerOptions": {
-        "allowSyntheticDefaultImports": true,
-        "experimentalDecorators": true
-    }
+    "extends": "../../../../../../tsconfig.dev.json"
 }
