+++ examples/experimental/test/end_to_end/http_server/internet_identity/src/frontend/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../../../../../../../tsconfig.dev.json",
+    "compilerOptions": {
+        "target": "ES2021",
+        "allowSyntheticDefaultImports": true,
+        "experimentalDecorators": true
+    }
+}
