+++ examples/experimental/test/end_to_end/http_server/internet_identity/tsconfig.json
@@ -1,8 +1,3 @@
 {
-    "extends": "../../../../../../tsconfig.dev.json",
-    "compilerOptions": {
-        "target": "ES2021",
-        "allowSyntheticDefaultImports": true,
-        "experimentalDecorators": true
-    }
+    "extends": "../../../../../../tsconfig.dev.json"
 }
